Delegate the Lucene95 codec and format implementation
