+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/segment/store/VectorIndexUtils.java
@@ -20,10 +20,12 @@ package org.apache.pinot.segment.local.segment.store;
 
 import java.io.File;
 import org.apache.commons.io.FileUtils;
+import org.apache.lucene.codecs.lucene95.Lucene95Codec;
+import org.apache.lucene.codecs.lucene95.Lucene95HnswVectorsFormat;
 import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.index.VectorSimilarityFunction;
-import org.apache.pinot.segment.local.segment.creator.impl.vector.lucene95.Lucene95Codec;
-import org.apache.pinot.segment.local.segment.creator.impl.vector.lucene95.Lucene95HnswVectorsFormat;
+import org.apache.pinot.segment.local.segment.creator.impl.vector.lucene95.HnswCodec;
+import org.apache.pinot.segment.local.segment.creator.impl.vector.lucene95.HnswVectorsFormat;
 import org.apache.pinot.segment.spi.V1Constants.Indexes;
 import org.apache.pinot.segment.spi.index.creator.VectorIndexConfig;
 
@@ -83,14 +85,14 @@ public class VectorIndexUtils {
     int beamWidth = Integer.parseInt(vectorIndexConfig.getProperties()
         .getOrDefault("beamWidth", String.valueOf(Lucene95HnswVectorsFormat.DEFAULT_BEAM_WIDTH)));
     int maxDimensions = Integer.parseInt(vectorIndexConfig.getProperties()
-        .getOrDefault("maxDimensions", String.valueOf(Lucene95HnswVectorsFormat.DEFAULT_MAX_DIMENSIONS)));
+        .getOrDefault("maxDimensions", String.valueOf(HnswVectorsFormat.DEFAULT_MAX_DIMENSIONS)));
 
-    Lucene95HnswVectorsFormat knnVectorsFormat =
-        new Lucene95HnswVectorsFormat(maxCon, beamWidth, maxDimensions);
+    HnswVectorsFormat knnVectorsFormat =
+        new HnswVectorsFormat(maxCon, beamWidth, maxDimensions);
 
     Lucene95Codec.Mode mode = Lucene95Codec.Mode.valueOf(vectorIndexConfig.getProperties()
         .getOrDefault("mode", Lucene95Codec.Mode.BEST_SPEED.name()));
-    indexWriterConfig.setCodec(new Lucene95Codec(mode, knnVectorsFormat));
+    indexWriterConfig.setCodec(new HnswCodec(mode, knnVectorsFormat));
     return indexWriterConfig;
   }
 }
